Mudando o valor do item para BigDecimal

+++ src/main/java/net/baade/Item.java
@@ -1,14 +1,16 @@
 package net.baade;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class Item {
-	public Item( String descricao, Float valor ) {
+	public Item( String descricao, BigDecimal valor ) {
 		this.descricao = descricao;
 		this.valor = valor;
 	}
 	private String descricao;
-	private Float valor;
+	private BigDecimal valor;
 
 }
